requirements.txt
Flask==2.3.3
torch==2.0.1
transformers==4.7.0
Pillow==9.3.0
